Bank Marketing Dataset
This dataset is taken from Kaggle https://www.kaggle.com/rouseguy/bankbalanced/version/1#
The data is related with direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be ('yes') or not ('no') subscribed.